backend-py/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sqlmodel>=0.0.8
psycopg2-binary>=2.9
alembic>=1.11
python-dotenv>=1.0
cloudinary>=1.34.0
